+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-11-06
+## [Unreleased] - 2023-11-11
 
 ### Added
+* Ability to specify list of paths to exclude from sitemap, via new input `exclude-paths`.
 
 ### Changed